+++ app.py
@@ -14,6 +14,7 @@ from lib.ui_mixins.recepciones_mixin import RecepcionesMixin
 from lib.ui_mixins.email_mixin import EmailMixin
 from lib.ui_mixins.clientes_mixin import ClientesMixin
 from lib.ui_mixins.dispositivos_qr_mixin import DispositivosQRMixin
+from lib.ui_mixins.trazabilidad_mixin import TrazabilidadMixin
 
 #try:
 #    locale.setlocale(locale.LC_ALL, 'es_ES.UTF-8')
@@ -254,7 +255,7 @@ class LoginApp(ctk.CTk):
 # 2. CLASE DE LA VENTANA PRINCIPAL DE LA APLICACIÓN
 # ----------------------------------------------------------------------
 
-class VentanaPrincipal(ctk.CTkToplevel, BusquedaMixin, DashboardMixin, TareasDashboardMixin, RmaListadoMixin, RmaEditorMixin, BackupsMixin, AdjuntosMixin, InformesMixin, AdminMixin, ArticulosMixin, RecepcionesMixin, EmailMixin, ClientesMixin, DispositivosQRMixin):
+class VentanaPrincipal(ctk.CTkToplevel, BusquedaMixin, DashboardMixin, TareasDashboardMixin, RmaListadoMixin, RmaEditorMixin, BackupsMixin, AdjuntosMixin, InformesMixin, AdminMixin, ArticulosMixin, RecepcionesMixin, EmailMixin, ClientesMixin, DispositivosQRMixin, TrazabilidadMixin):
     """Ventana principal que gestiona el listado, creación y edición de RMAs."""
     
     # Opciones predefinidas para desplegables